rendu/resources/requirements.txt
unidecode
filelock
regex
gunicorn
fastspellchecker
pillow
python-dotenv
flask
pytest
numpy
